Move a function to a more appropriate location
`pulp_smash.tests.pulp3.utils.get_content_unit_paths` only works for
file repositories. Move it into
`pulp_smash.tests.pulp3.file.api_v3.utils`.

Fix: https://github.com/PulpQE/pulp-smash/issues/979
