+++ pulp_smash/tests/pulp3/file/api_v3/test_download_content.py
@@ -14,17 +14,13 @@ from pulp_smash.tests.pulp3.constants import (
     REPO_PATH,
 )
 from pulp_smash.tests.pulp3.file.api_v3.utils import (
-    gen_remote,
     gen_publisher,
+    gen_remote,
+    get_content_unit_paths,
 )
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
 from pulp_smash.tests.pulp3.pulpcore.utils import gen_distribution, gen_repo
-from pulp_smash.tests.pulp3.utils import (
-    get_auth,
-    get_content_unit_paths,
-    publish_repo,
-    sync_repo,
-)
+from pulp_smash.tests.pulp3.utils import get_auth, publish_repo, sync_repo
 
 
 class DownloadContentTestCase(unittest.TestCase, utils.SmokeTest):
